re-org compile mod
