+++ mlx-rs/src/transforms/compile/compile_updatable.rs
@@ -1,0 +1,10 @@
+
+/// # Generics:
+/// 
+/// - `U`: The type of the updatable
+/// - `A`: The type of the array arguments
+/// - `O`: The type of the output
+/// - `E`: The type of the error
+pub trait CompileUpdatable<U, A, O, E> {
+    // fn compile_updatable(self, updatable: U)
+}
